Moved all main project files into new folder \project
